Move confirmationEmail definition to forms system (#18878)
* Add confirmationEmail definition

* Use platform confirmationEmail definition
