+++ src/applications/caregivers/config/chapters/secondaryTwo/secondaryTwoContactInfo.js
@@ -1,4 +1,5 @@
 import fullSchema from 'vets-json-schema/dist/10-10CG-schema.json';
+import confirmationEmailUI from 'platform/forms-system/src/js/definitions/confirmationEmail';
 import { SecondaryCaregiverInfo } from 'applications/caregivers/components/AdditionalInfo';
 import { secondaryTwoFields } from 'applications/caregivers/definitions/constants';
 import { secondaryTwoInputLabel } from 'applications/caregivers/definitions/UIDefinitions/caregiverUI';
@@ -8,7 +9,6 @@ import {
   alternativePhoneNumberUI,
   primaryPhoneNumberUI,
   addressWithoutCountryUI,
-  confirmationEmailUI,
 } from 'applications/caregivers/definitions/UIDefinitions/sharedUI';
 
 const { secondaryCaregiverTwo } = fullSchema.properties;
