+++ src/applications/caregivers/config/chapters/veteran/vetContactInfo.js
@@ -1,4 +1,5 @@
 import fullSchema from 'vets-json-schema/dist/10-10CG-schema.json';
+import confirmationEmailUI from 'platform/forms-system/src/js/definitions/confirmationEmail';
 import { VetInfo } from 'applications/caregivers/components/AdditionalInfo';
 import { veteranFields } from 'applications/caregivers/definitions/constants';
 import { vetInputLabel } from 'applications/caregivers/definitions/UIDefinitions/veteranUI';
@@ -7,7 +8,6 @@ import {
   emailUI,
   alternativePhoneNumberUI,
   primaryPhoneNumberUI,
-  confirmationEmailUI,
   addressWithoutCountryUI,
 } from 'applications/caregivers/definitions/UIDefinitions/sharedUI';
