+++ src/applications/caregivers/definitions/UIDefinitions/sharedUI.js
@@ -72,26 +72,6 @@ export const ssnUI = label => ({
   ],
 });
 
-export const confirmationEmailUI = (label, dataConstant) => ({
-  'ui:title': `Re-enter ${label}  email address`,
-  'ui:widget': 'email',
-  'ui:required': formData => !!formData[dataConstant],
-  'ui:validations': [
-    {
-      validator: (errors, fieldData, formData) => {
-        const doesEmailMatch = () =>
-          formData[dataConstant] === formData[`view:${dataConstant}`];
-
-        if (!doesEmailMatch()) {
-          errors.addError(
-            'This email does not match your previously entered email',
-          );
-        }
-      },
-    },
-  ],
-});
-
 export const addressWithoutCountryUI = label => ({
   'ui:title': ' ',
   'ui:order': ['street', 'street2', 'city', 'state', 'postalCode'],
